src: use `void` as value type for `unique_files`
hashmaps already store the unhashed key; they use these to check if a key
is actually equivalent on hash-collision.

So we can avoid storing the unhashed keys manually by setting the value
type to `void`.
Additionally remove all self-implemented logic related to hash-collisions.
Also remove the `FileList` type as it now goes unused.

Due to the removal of the hash-collision logic and our mismanaging of
unhashed keys, the program can no longer diferentiate keys in-case of
hash-collisions. This will be addressed in a future commit.