src: use `sys_alloc` for HashMaps
When a HashMap calls grow() it will create a new allocation and copy all
the values into it; then it calls free() on the old allocation.
ArenaAllocator's free() only works if the object it is called on is the
most recent allocation; so due to HashMaps new copy always being the
latest, free() never does anything, and the arena will just keep growing.

So switch the `sys_alloc` to prevent excessive memory usage.