perf: only store file if two could be held at once
We need to be capable of storing 2 copies of a file at one given moment
to see if they're duplicates of eachother; if we can't, it's an error.
So, introduce `max_file_size` as an upperlimit to stop loading files we
could never store two of.

The current version of `max_file_size` will only be effective on 32-bit
or other small pointer systems. In the future `max_file_size`could be
based of other memory we know to *always* be taken; potentially making
it useful for 64-bit systems.