feat: track unique files and mark duplicates
Previously didn't actually do this,
as the stored key values would be discarded as the arena was reset.

Now we use a context that reads and compares files in chunks.
This would effectively remove maximum filesize, if not for the fact that
we store the size to speed up comparison.